Ft gratefuls journals single

+++ app/controllers/gratefuls.py
@@ -3,8 +3,9 @@ from flask_restful import Resource, request
 
 from app.schemas import GratefulSchema
 from app.models.gratefuls import Grateful
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt_claims
 from flask_bcrypt import Bcrypt
+from app.helpers.admin import is_owner_or_admin
 
 
 class GratefulsView(Resource):
@@ -56,11 +57,31 @@ class GratefulsView(Resource):
 
 class GratefulsDetailView(Resource):
     @jwt_required
+    def get(self, grateful_id):
+        """
+        Getting a particular grateful
+        """
+        grateful_schema = GratefulSchema(many=False)
+
+        grateful = Grateful.get_by_id(grateful_id)
+
+        if not grateful:
+            return dict(status='fail', message=f'No grateful data'), 404
+        
+        grateful_data, errors = grateful_schema.dumps(grateful)
+
+        if errors:
+            return dict(status="fail", message="Internal Server Error"), 500
+        
+        return dict(status="success", data=dict(grateful=json.loads(grateful_data)))
+
+    @jwt_required
     def patch(self, grateful_id):
         """
         Update user grateful
         """
-
+        current_user_id = get_jwt_identity()
+        current_user_roles = get_jwt_claims()['roles'] 
         grateful_schema = GratefulSchema(partial=True)
 
         update_data = request.get_json()
@@ -77,6 +98,9 @@ class GratefulsDetailView(Resource):
                 status="fail",
                 message=f"Grateful with id {grateful_id} not found"
             ), 404
+        
+        if not is_owner_or_admin(grateful, current_user_id, current_user_roles):
+            return dict(status='fail', message='unauthorised'), 403
 
         updated_grateful = Grateful.update(grateful, **validated_update_data)
 
@@ -93,7 +117,8 @@ class GratefulsDetailView(Resource):
         """
         Delete a user grateful
         """
-
+        current_user_id = get_jwt_identity()
+        current_user_roles = get_jwt_claims()['roles']  
         grateful = Grateful.get_by_id(grateful_id)
 
         if not grateful:
@@ -101,6 +126,9 @@ class GratefulsDetailView(Resource):
                 status="fail",
                 message=f"Grateful with id {grateful_id} not found"
             ), 404
+        
+        if not is_owner_or_admin(grateful, current_user_id, current_user_roles):
+            return dict(status='fail', message='unauthorised'), 403
 
         deleted_grateful = grateful.delete()
